+++ tests/Actions/Settings/ResponseDelete.php
@@ -18,19 +18,19 @@ class Actions_Settings_ResponseDelete extends ActionTestCase
         $this->assertInstanceOf('rcmail_action', $action);
         $this->assertTrue($action->checks());
 
-        $responses = [
-            [
-                'name' => 'test',
-                'text' => '****',
-                'format' => 'text',
-                'key' => '1234',
-            ]
-        ];
-
         $rcmail = rcmail::get_instance();
-        $rcmail->user->save_prefs(['compose_responses' => $responses]);
+        $rcmail->user->save_prefs(['compose_responses_static' => []]);
 
-        $_POST = ['_key' => '1234'];
+        self::initDB('responses');
+
+        $responses = $rcmail->get_compose_responses();
+
+        $this->assertCount(2, $responses);
+
+        $rid = $responses[0]['id'];
+
+        // Test successful request
+        $_POST = ['_id' => $rid];
 
         $this->runAndAssert($action, OutputJsonMock::E_EXIT);
 
@@ -39,8 +39,24 @@ class Actions_Settings_ResponseDelete extends ActionTestCase
         $this->assertSame(['Content-Type: application/json; charset=UTF-8'], $output->headers);
         $this->assertSame('delete-response', $result['action']);
         $this->assertTrue(strpos($result['exec'], 'this.display_message("Successfully deleted.","confirmation");') !== false);
-        $this->assertTrue(strpos($result['exec'], 'this.remove_response("1234")') !== false);
+        $this->assertTrue(strpos($result['exec'], 'this.remove_response("' . $rid .'")') !== false);
 
-        $this->assertSame([], $rcmail->get_compose_responses());
+        $responses = $rcmail->get_compose_responses();
+
+        $this->assertCount(1, $responses);
+        $this->assertTrue($responses[0]['id'] != $rid);
+
+        // Test error
+        $_POST = ['_id' => 'unknown'];
+
+        $this->runAndAssert($action, OutputJsonMock::E_EXIT);
+
+        $result = $output->getOutput();
+
+        $this->assertSame(['Content-Type: application/json; charset=UTF-8'], $output->headers);
+        $this->assertSame('delete-response', $result['action']);
+        $this->assertTrue(strpos($result['exec'], 'this.display_message("An error occurred while saving.","error"') !== false);
+
+        $this->assertCount(1, $rcmail->get_compose_responses());
     }
 }