+++ tests/Actions/Settings/ResponseEdit.php
@@ -18,19 +18,34 @@ class Actions_Settings_ResponseEdit extends ActionTestCase
         $this->assertInstanceOf('rcmail_action', $action);
         $this->assertTrue($action->checks());
 
-        $responses = [
-            [
-                'name' => 'test',
-                'text' => '****',
-                'format' => 'text',
-                'key' => '1234',
-            ]
-        ];
-
         $rcmail = rcmail::get_instance();
-        $rcmail->user->save_prefs(['compose_responses' => $responses]);
+        $rcmail->user->save_prefs([
+            'compose_responses_static' => [
+                ['name' => 'static 1', 'text' => 'Static Response One'],
+            ]
+        ]);
 
-        $_GET = ['_key' => '1234'];
+        self::initDB('responses');
+
+        $responses = $rcmail->get_compose_responses();
+
+        // Test read-only response
+        $_GET = ['_id' => $responses[0]['id']];
+
+        $this->runAndAssert($action, OutputHtmlMock::E_EXIT);
+
+        $result = $output->getOutput();
+
+        $this->assertSame('responseedit', $output->template);
+        $this->assertSame('Edit response', $output->getProperty('pagetitle'));
+        $this->assertSame(true, $output->get_env('readonly'));
+        $this->assertTrue(stripos($result, "<!DOCTYPE html>") === 0);
+        $this->assertTrue(strpos($result, "rcmail.gui_object('editform', 'form')") !== false);
+        $this->assertTrue(strpos($result, "tinymce.min.js") !== false);
+        $this->assertTrue(strpos($result, "Static Response One</textarea>") !== false);
+
+        // Test writable response
+        $_GET = ['_id' => $responses[2]['id']];
 
         $this->runAndAssert($action, OutputHtmlMock::E_EXIT);
 
@@ -39,9 +54,7 @@ class Actions_Settings_ResponseEdit extends ActionTestCase
         $this->assertSame('responseedit', $output->template);
         $this->assertSame('Edit response', $output->getProperty('pagetitle'));
         $this->assertSame(false, $output->get_env('readonly'));
-        $this->assertTrue(stripos($result, "<!DOCTYPE html>") === 0);
-        $this->assertTrue(strpos($result, "rcmail.gui_object('editform', 'form')") !== false);
-        $this->assertTrue(strpos($result, "****") !== false);
+        $this->assertTrue(strpos($result, "test response 2&lt;/b&gt;&lt;/p&gt;</textarea>") !== false);
     }
 
     /**